refactor: Muda double e double* para sample_t e signal_t nos arquivos de wavelet transform e thresholding

+++ headers/signal_base.h
@@ -1,7 +1,9 @@
 #ifndef __SIGNAL_BASE_H__
 #define __SIGNAL_BASE_H__
 
-typedef double sample_t;
+#include <stdint.h>
+
+typedef int sample_t;
 typedef sample_t* signal_t;
 
 int sample_compare_function(const void* a, const void* b);
